Rework Bot Govern to maximize patronage gain
In the case where the Bot can govern 2 spaces but is low on
us Aid.  We ensure that it adds add in the lower population
space first so that it can then add to patronage in the higher
population space.
We also ensure that if Mandate of Heaven (unshaded) is in effect,
that we attempt avoid shifting from Passive Support to Neutral
if possible.