Fix bug in Bot sweep code
If we moved Troops to a space from multiple origins, then we were adding the
sweep destination so the `moveDestinations` list multiple times.
This was causing the code to activate guerrillas once for each instance
of the destination in the list.  We now prevent the destination from being
added to the list if it is already present.