Fix issue with Bot movement and moveable pieces
When checking movable pieces in a space, do not include those that
have already moved into the space.

Fix implementation of NVA march.
One of the priorities states:
Select space adjacent to most NVA troops yet to march (if trail is 0-3)

We were selecting this base on all moveable troops in all spaces adjacent to the
target space.  But according to the designer, Bruce Mansfield, we should only consider
the one adjacent space with the most moveable pieces.
(See this BGG post: https://boardgamegeek.com/thread/2969461/article/41151414#41151414)

I think this is to make it easier to carry out the Bot instruction.
So I changed the code to implement it in this way.