Don't stop Sweep early if user has SA
The program was ending the sweep command in the case where no
guerrillas could be activated in any of the Sweep spaces.
But in the case where the user may perform a Special Activity,
we must wait to see if the user chooses to Air Lift troop into one
of the Sweep spaces.