Rework code to show game state differences
When a game turn is aborted or the turn is rolled back,
we show the necessary changes that need to be made to the game board.
The display now first show all pieces that must be removed from each space and
from casualties and out of play so that the surplus pieces are first added to
available.  Then we show which pieces should be added to each space and to
casualties and out of play.  This should make it easier to update the physical
board state.