Show menu when asking about pivotal events
Fixes #20

We now show all factions that are eligible to play pivotal events.
We show which Bot faction has elected to play their's (if any)
And we show which factions have declined.

Then we show a menu to the Human factions (in order of precedence) asking
if they wish to play their pivotal event with options:
1. Yes
2. No
3. Show Coup status