Show Bot event intentions
Fixes #22

When showing the current card and on deck card we now show for each Bot faction
the Bot's intention for the event based on the Trung Event chart.
(Critical, Performed, Ignored)
The user can choose between verbose display, concise display, or to not display it all.

Also changed the NotExecuted bot priority to Ignored.